component clean up
