--- src/__swaps__/screens/Swap.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export function SwapScreen() {
-  return <>Swaps v2</>;
-}
-
-export default SwapScreen;
