+++ src/navigation/Routes.ios.tsx
@@ -97,6 +97,7 @@ import { PointsProfileProvider } from '@/screens/points/contexts/PointsProfileCo
 import AppIconUnlockSheet from '@/screens/AppIconUnlockSheet';
 import { SwapScreen } from '@/__swaps__/screens/Swap/Swap';
 import { useRemoteConfig } from '@/model/remoteConfig';
+import { SwapProvider } from '@/__swaps__/screens/Swap/providers/swap-provider';
 
 type StackNavigatorParams = {
   [Routes.SEND_SHEET]: unknown;
@@ -134,6 +135,16 @@ function MainStack() {
     <Stack.Navigator initialRouteName={Routes.MAIN_NAVIGATOR_WRAPPER} {...stackNavigationConfig} screenOptions={defaultScreenStackOptions}>
       <Stack.Screen component={MainNavigator} name={Routes.MAIN_NAVIGATOR_WRAPPER} />
     </Stack.Navigator>
+  );
+}
+
+function SwapsNavigator() {
+  return (
+    <SwapProvider>
+      <NativeStack.Navigator {...nativeStackConfig} initialRouteName={Routes.SWAP}>
+        <NativeStack.Screen component={SwapScreen} name={Routes.SWAP} {...swapConfig} />
+      </NativeStack.Navigator>
+    </SwapProvider>
   );
 }
 
@@ -290,7 +301,7 @@ function NativeStackNavigator() {
       <NativeStack.Screen component={ConsoleSheet} name={Routes.CONSOLE_SHEET} {...consoleSheetConfig} />
       <NativeStack.Screen component={AppIconUnlockSheet} name={Routes.APP_ICON_UNLOCK_SHEET} {...appIconUnlockSheetConfig} />
 
-      {swapsV2Enabled && <NativeStack.Screen component={SwapScreen} name={Routes.SWAP} {...swapConfig} />}
+      {swapsV2Enabled && <NativeStack.Screen component={SwapsNavigator} name={Routes.SWAP_NAVIGATOR} {...swapConfig} />}
     </NativeStack.Navigator>
   );
 }
