+++ src/navigation/routesNames.ts
@@ -19,7 +19,10 @@ const Routes = {
   CUSTOM_GAS_SHEET: 'CustomGasSheet',
   DAPP_BROWSER_SCREEN: 'DappBrowserScreen',
   DAPP_BROWSER: 'DappBrowser',
+  SWAP_NAVIGATOR: 'SwapNavigator',
   SWAP: 'Swap',
+  SWAP_REVIEW: 'SwapReview',
+  SWAP_GAS: 'SwapGas',
   DIAGNOSTICS_SHEET: 'DiagnosticsSheet',
   DISCOVER_SCREEN: 'DiscoverScreen',
   ENS_ADDITIONAL_RECORDS_SHEET: 'ENSAdditionalRecordsSheet',
